+++ crates/vertigo/src/driver_module/src_js/api/command/dom/map_nodes.ts
@@ -1,5 +1,3 @@
-import { trySaveMetaData } from "../../metadata";
-
 type NodeType = Element | Comment | Text;
 export class MapNodes {
     private data: Map<number, NodeType>;
@@ -141,7 +139,6 @@ export class MapNodes {
         }
 
         for (const node of initNodes) {
-            trySaveMetaData(node);
             node.remove();
         }
     }